rename config_example.py to config.py in example directory
